third_party/blink/web_tests/fast/loader/hashchange-event-properties-expected.txt

Tests that hashchange events have the expected newURL and oldURL properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

hashchange fired with oldURL hash "[none]" and newURL hash "#state1"
hashchange fired with oldURL hash "#state1" and newURL hash "#state2"
hashchange fired with oldURL hash "#state2" and newURL hash "#state1"
hashchange fired with oldURL hash "#state1" and newURL hash "[none]"
PASS successfullyParsed is true

TEST COMPLETE
